+++ FBSDKLoginKit/FBSDKLoginKit/Internal/FBSDKLoginManager+Internal.h
@@ -32,7 +32,7 @@ NS_SWIFT_NAME(BrowserLoginSuccessBlock);
 @property (nullable, nonatomic) FBSDKLoginConfiguration *configuration;
 @property (nonatomic) id<FBSDKKeychainStore> keychainStore;
 @property (nonatomic) Class<FBSDKAccessTokenProviding> accessTokenWallet;
-@property (nonatomic) Class<FBSDKAuthenticationTokenProviding, FBSDKAuthenticationTokenSetting> authenticationToken;
+@property (nonatomic) Class<FBSDKAuthenticationTokenProviding> authenticationToken;
 @property (nonatomic) Class<FBSDKProfileProviding> profile;
 @property (nonatomic) id<FBSDKURLHosting, FBSDKAppURLSchemeProviding, FBSDKAppAvailabilityChecker> internalUtility;
 @property (nonatomic) id<FBSDKURLOpener> urlOpener;
@@ -52,7 +52,7 @@ NS_SWIFT_NAME(BrowserLoginSuccessBlock);
 - (instancetype)initWithInternalUtility:(id<FBSDKURLHosting, FBSDKAppURLSchemeProviding, FBSDKAppAvailabilityChecker>)internalUtility
                    keychainStoreFactory:(id<FBSDKKeychainStoreProviding>)keychainStoreFactory
                       accessTokenWallet:(Class<FBSDKAccessTokenProviding>)accessTokenWallet
-                    authenticationToken:(Class<FBSDKAuthenticationTokenProviding, FBSDKAuthenticationTokenSetting>)authenticationToken
+                    authenticationToken:(Class<FBSDKAuthenticationTokenProviding>)authenticationToken
                                 profile:(Class<FBSDKProfileProviding>)profile
                               urlOpener:(id<FBSDKURLOpener>)urlOpener
                                settings:(id<FBSDKSettings>)settings